% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{arcpbf-package}
\alias{arcpbf}
\alias{arcpbf-package}
\title{arcpbf: Process ArcGIS Protocol Buffer FeatureCollections}
\description{
Fast processing of ArcGIS FeatureCollection protocol buffers in R. It is designed to work seamlessly with 'httr2' and integrates with 'sf'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.esri.com/arcpbf/}
  \item \url{https://github.com/R-ArcGIS/arcpbf}
  \item Report bugs at \url{https://github.com/R-ArcGIS/arcpbf/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
