% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{multi_resp_process}
\alias{multi_resp_process}
\title{Process a list of httr2 responses}
\usage{
multi_resp_process(resps)
}
\arguments{
\item{resps}{a list of \code{httr2_response} objects such as
created by \code{httr2::multi_req_perform()}}
}
\value{
A list where each element is a processed FeatureCollection PBF.
}
\description{
When running multiple requests in parallel using \code{\link[httr2:multi_req_perform]{httr2::multi_req_perform()}}
the responses are returned as a list of responses. \code{multi_resp_process()} processes
the responses in a vectorized manner.
}
\details{
If a response is not 200 status code or does not have the appropriate \code{"application/x-protobuf"}
content type, the result will be \code{NULL}.

The results of this \emph{are not} post processed. Post processing can be
applied to the resultant list object using \code{post_process_pbf()}.
See the example for a fully worked example including post-processing.
}
\examples{
if (rlang::is_installed("httr2") && interactive()) {
    url <- "https://services.arcgis.com/P3ePLMYs2RVChkJx/arcgis/rest/services/ACS_Population_by_Race_and_Hispanic_Origin_Boundaries/FeatureServer/2/query?where=1=1&outFields=*&f=pbf&token="
    reqs <- replicate(5, httr2::request(url), simplify = FALSE)
    resps <- httr2::multi_req_perform(reqs)
    pbfs <- multi_resp_process(resps)
    post_process_pbf(pbfs)
}
}
\seealso{
Other httr2: 
\code{\link{resp_body_pbf}()}
}
\concept{httr2}
