% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{open_pbf}
\alias{open_pbf}
\title{Read a pbf file as a raw vector}
\usage{
open_pbf(path)
}
\arguments{
\item{path}{the path to the \code{.pbf} file.}
}
\value{
a raw vector
}
\description{
Read a pbf file as a raw vector
}
\examples{
count_fp <- system.file("count.pbf", package = "arcpbf")
oid_fp <- system.file("ids.pbf", package = "arcpbf")
tbl_fp <- system.file("small-table.pbf", package = "arcpbf")
fc_fp <- system.file("small-points.pbf", package = "arcpbf")
count_raw <- open_pbf(count_fp)
oid_raw <- open_pbf(oid_fp)
tbl_raw <- open_pbf(tbl_fp)
fc_raw <- open_pbf(fc_fp)
}
