% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-body-pbf.R
\name{resp_body_pbf}
\alias{resp_body_pbf}
\title{Extract body from response}
\usage{
resp_body_pbf(resp, post_process = TRUE, use_sf = TRUE)
}
\arguments{
\item{resp}{A response object.}

\item{post_process}{default \code{TRUE}. Apply \code{post_process_pbf()} to the
pbf body.}

\item{use_sf}{default \code{TRUE}. Whether or not to return an \code{sf} object.}
}
\description{
Responses of type \code{application/x-protobuf} are automatically
processed using \code{process_pbf()} with optional post-processing
applied.
}
\details{
This function assumes the body is an Esri FeatureCollection
protobuf. It does not support Knowledge Graph protobufs.
}
\examples{

if (rlang::is_installed("httr2")) {
  url <- "https://services.arcgis.com/P3ePLMYs2RVChkJx/arcgis/rest/services/ACS_Population_by_Race_and_Hispanic_Origin_Boundaries/FeatureServer/2/query?where=1=1&outFields=objectid&resultRecordCount=10&f=pbf&token="

  req <- httr2::request(url)
  resp <- httr2::req_perform(req)
  resp_body_pbf(resp)
}
}
\seealso{
Other httr2: 
\code{\link{multi_resp_process}()}
}
\concept{httr2}
