% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{multi_resp_process}
\alias{multi_resp_process}
\title{Process a list of httr2 responses}
\usage{
multi_resp_process(resps)
}
\arguments{
\item{resps}{a list of \code{httr2_response} objects such as
created by \code{httr2::multi_req_perform()}}
}
\description{
Process a list of httr2 responses
}
