% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process_pbf}
\alias{post_process_pbf}
\title{Post process pbf results}
\usage{
post_process_pbf(x, use_sf = TRUE)
}
\arguments{
\item{x}{an object as returned by \code{process_pbf()} or \code{read_pbf()}}

\item{use_sf}{default \code{TRUE}. Whether or not to return an \code{sf} object.
If \code{FALSE}, a \code{data.frame} is returned with the spatial reference stored
in the \code{crs} attribute.}
}
\value{
An object of class \code{data.frame} or \code{sf} when \code{use_sf = TRUE}.
}
\description{
Post process pbf results
}
