% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{process_pbf}
\alias{process_pbf}
\title{Process a FeatureCollection PBF}
\usage{
process_pbf(proto)
}
\arguments{
\item{proto}{either a raw vector or a list of raw vectors.}
}
\value{
\code{NULL} if the object is not a raw vector or a list of them.
If \code{proto} is a raw vector and the FeatureCollection pbf has a
geometry, then a name list with three elements \code{attributes},
\code{sr}, and \code{geometry} is returned. If there is no geometry,
then a \code{data.frame} is returned. If FeatureCollection is a
\code{CountResult} then a scalar integer is returned.
}
\description{
Process a pbf from a raw vector or a list of raw vectors.
}
